Challenge 2
I removed validation for expired passwords. Added universities page after login that fetches the data and inserts into a table. Created class for API and University Model. As for login improvements i would suggest making a separate class/controller that would be responsible for redirecting and logging in for every page.